requirements.txt
attrs==19.1.0
boto3==1.9.185
dj-inmemorystorage==1.4.0
Django==1.11.21
django-admin-sortable==2.0.12
django-allauth==0.39.1
django-appconf==1.0.1
django-countries==3.0.1
django-debug-toolbar==1.11
django-debug-toolbar-template-timings==0.9
django-extensions==2.1.9
django-filter==0.9.2
django-flat-theme==1.1.3
django-hosts==3.0
django-image-cropping==1.0.3
django-lockdown==1.2
django-logentry-admin==0.1.5
django-markdownx==2.0.28
django-modeltranslation==0.13.1
django-nested-admin==3.2.3
django-nose==1.4.5
django-storages==1.7.1
django-watson==1.5.2
django_select2==6.3.1
djangorestframework==3.9.4
easy-thumbnails==2.6.0
gunicorn==19.7.1
model-mommy==1.3.1
nose==1.3.7
oauthlib==0.7.2
opbeat==3.1.2
Pillow==3.4.2
psycopg2==2.8.3
python-openid==2.2.5
requests==2.4.3
requests-oauthlib==0.4.2
six==1.9.0
sqlparse==0.3.0
unicodecsv==0.9.4
wheel==0.29.0
